+++ tests/testthat/test_miscUtils.R
@@ -72,3 +72,20 @@ test_that("lagratio() calculations",{
   expect_equal(lagratio(1:10,3,2,direction="forward"),res3r)
 })
 
+test_that("changesPos() error messages",{
+  ## check error messages
+  expect_error(changesPos(numeric(0)),"length")
+  expect_error(changesPos(1,include.first=FALSE),"include.first")
+  expect_error(changesPos(matrix(1:4)),"vector")
+  expect_error(changesPos(data.frame(x=1:4)),"vector")
+})
+
+test_that("changesPos() calculations",{
+  expect_equal(changesPos(1:4),1:4)
+  expect_equal(changesPos(c(1:2,1:2)),1:4)
+  expect_equal(changesPos(c(1,2,2,1)),c(1,2,4))
+  expect_equal(changesPos(c(1,2,2,1),include.first=FALSE),c(2,4))
+  expect_equal(changesPos(c(1,1,1,1)),1)
+  expect_equal(changesPos(c(1,1,1,1),include.first=FALSE),numeric(0))
+})  
+