+++ NEWS.md
@@ -2,6 +2,7 @@
 * Added travis-ci integration.
 * Added coveralls integration.
 * Added `importFrom` for `mapvalues()` from `plyr`.
+* `changesPos()`: Modified.  Added some checks with error messages.  Added suite of tests.
 * `dunnTest()`: Modified.  Change class type from `DunnTest` to `dunnTest` to eliminate conflict with `DunnTest()` in `DescTools` package.  Thanks to Sal Mangiafico for pointing out this conflict.
 * `lagratio()`: Modified.  Corrected incorrect explanation of `differences=` in help file.  Added `recursion=` and `direction=`.  Added suite of tests.
 * `mapvalues()`: Deleted.  Deleted the function from `FSA` but imported it from `plyr` and then exported it from `FSA` so that it would be available to fisheries users without having to load `plyr`.